Move lib/blog to app/blog/lib
