+++ app/blog/lib/blogPost.ts
