+++ app/blog/lib/getPureCloudinaryPath.ts
