+++ app/blog/page.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import Title from "@/components/Title";
-import {getSortedPostsData} from "@/lib/blog/load";
+import {getSortedPostsData} from "@blog/lib/load";
 
 import styles from '../../styles/blog/blog.module.scss';
 
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faStar} from "@fortawesome/free-solid-svg-icons";
 import ArticleCard, {ArticleGrid} from "@/components/blog/ArticleCard";
-import BlogPost from "@/lib/blog/blogPost";
+import BlogPost from "@blog/lib/blogPost";
 import {Metadata} from "next";
 import MainWrapper from "@/components/MainWrapper";
