+++ app/blog/renderer/BlogMarkdown.tsx
@@ -3,13 +3,13 @@ import React, {CSSProperties} from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import rehypeRaw from "rehype-raw";
-import {BlogImageData} from "@/lib/blog/imagePropsFetcher";
+import {BlogImageData} from "@blog/lib/imagePropsFetcher";
 import PageNavigation from "@/components/blog/PageNavigation";
 import Block from "@/components/Block";
 
 import RendererProvider from "./RendererProvider";
 import ArticleRendererFromContext from "./ArticleRenderer";
-import BlogPost from "@/lib/blog/blogPost";
+import BlogPost from "@blog/lib/blogPost";
 
 export const parseInlineMarkdown = (markdown: string) => {
   const comp = {
