+++ components/blog/ArticleCard/index.tsx
@@ -5,8 +5,8 @@ import dayjs from "dayjs";
 import Tag from "../Tag";
 import {ParseWithBudouX} from "@/lib/wordSplit";
 import React from "react";
-import {getPureCloudinaryPath} from "@/lib/blog/getPureCloudinaryPath";
-import BlogPost from "@/lib/blog/blogPost";
+import {getPureCloudinaryPath} from "@blog/lib/getPureCloudinaryPath";
+import BlogPost from "@blog/lib/blogPost";
 import Balancer from "react-wrap-balancer";
 import CldImageWrapper from "../../utils/CldImageWrapper";
 import BlockLink from "../../BlockLink";
