+++ components/blog/article-parts/TwitterArchive/index.tsx
@@ -2,7 +2,7 @@ import styles from "./index.module.scss";
 import React from "react";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faDove} from "@fortawesome/free-solid-svg-icons";
-import getOtakuColor from "@/lib/blog/otakuColors";
+import getOtakuColor from "@blog/lib/otakuColors";
 import {ArticleParts} from "../../ArticleParts";
 
 const TwitterArchive: ArticleParts = content => {
