+++ components/utils/CldImageWrapper.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import {CldImage, CldImageProps} from "next-cloudinary";
 import Image, {ImageProps} from "next/image";
-import cloudinaryLoader from "@/lib/blog/cloudinaryLoader";
+import cloudinaryLoader from "@blog/lib/cloudinaryLoader";
 
 export default function CldImageWrapper(
   props: ImageProps
