+++ tsconfig.json
@@ -22,7 +22,10 @@
     "paths": {
       "@/*": [
         "*"
-      ]
+      ],
+      "@blog/*": [
+        "app/blog/*"
+      ],
     },
     "plugins": [
       {
