+++ watchMarkdown.js
@@ -8,7 +8,7 @@ const chokidar = require('chokidar');
 
 let isUpdating = false
 const postsDir = './posts'
-const scriptDir = './lib/blog/fileWatch.ts'
+const scriptDir = './app/blog/lib/fileWatch.ts'
 const originalFileContent = fs.readFileSync(scriptDir, 'utf-8')
 
 const handler = () => {
